Extract resolvable interface to its own file
