+++ packages/alfa-css/src/value/numeric/percentage.ts
@@ -8,10 +8,12 @@ import {
   Percentage as BasePercentage,
 } from "../../calculation/numeric";
 
+import { Token } from "../../syntax";
+
+import { Resolvable } from "../resolvable";
 import { Value } from "../value";
 
 import { Numeric } from "./numeric";
-import { Token } from "../../syntax";
 
 const { either, map } = Parser;
 
@@ -147,7 +149,7 @@ export namespace Percentage {
   export interface Resolver<
     R extends BaseNumeric.Type,
     T extends Numeric.Fixed<R>
-  > extends Value.Resolver {
+  > {
     percentageBase: T;
   }
 
@@ -155,7 +157,7 @@ export namespace Percentage {
     CALC extends boolean = boolean,
     R extends BaseNumeric.Type = BaseNumeric.Type
   > extends Value<"percentage", CALC, "percentage" | R>,
-      Value.Resolvable<"percentage" | R, Resolver<R, Numeric.Fixed<R>>> {
+      Resolvable<"percentage" | R, Resolver<R, Numeric.Fixed<R>>> {
     hasCalculation(): this is Calculated<R>;
     resolve<T extends Numeric.Fixed<R>>(
       resolver?: Resolver<R, T>
